OBMIN-466 Move unit test to another class
